add const.js and utils.js
