+++ src/helpers/const.js
@@ -1,0 +1,7 @@
+export const GENRES = [
+  'Musical',
+  'Western',
+  'Drama',
+  'Comedy',
+  'Cartoon',
+];
