+++ src/mock/film.js
@@ -1,17 +1,7 @@
-// import dayjs from 'dayjs';
+import { getRandomInteger, getFalseOrTrue, generateValue, generateDescription } from '../helpers/utils'; // это не тут должно быть?
+import { GENRES } from '../helpers/const';
 
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-const getFalseOrTrue = () => {
-  return getRandomInteger(0,1) === 0 ? false : true;
-};
-
-const descriptions = [
+export const description = [
   'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
   'Cras aliquet varius magna, non porta ligula feugiat eget.',
   'Fusce tristique felis at fermentum pharetra.',
@@ -25,7 +15,7 @@ const descriptions = [
   'In rutrum ac purus sit amet tempus.',
 ];
 
-const titles = [
+export const titles = [
   'made-for-each-other',
   'popeye-meets-sinbad',
   'sagebrush-trail',
@@ -35,7 +25,7 @@ const titles = [
   'the-man-with-the-golden-arm',
 ];
 
-const posters = [
+export const posters = [
   './images/posters/made-for-each-other.png',
   './images/posters/popeye-meets-sinbad.png',
   './images/posters/sagebrush-trail.jpg',
@@ -45,31 +35,17 @@ const posters = [
   './images/posters/the-man-with-the-golden-arm.jpg',
 ];
 
-const genres = [
-  'Musical',
-  'Western',
-  'Drama',
-  'Comedy',
-  'Cartoon',
-];
 
-const generateValue = (inputArr) => {
-  const randomIndex = getRandomInteger(0, inputArr.length - 1);
-  return inputArr[randomIndex];
-};
 
 export const generateFilm = () => {
-
-  const description = new Array(getRandomInteger(1, 5)).fill().map(() => generateValue(descriptions));
-
   return {
     title: generateValue(titles),
     poster: generateValue(posters),
-    description,
+    description: generateDescription(description),
     raiting: `${getRandomInteger(1, 9)}.${getRandomInteger(1, 9)}`,
     bornYear: getRandomInteger(1930, 2000),
     duration: `${getRandomInteger(1, 3)}h.${getRandomInteger(1, 59)}m`,
-    genre: generateValue(genres),
+    genre: generateValue(GENRES),
     isWatchList: getFalseOrTrue(),
     isWatched: getFalseOrTrue(),
     isFavorite: getFalseOrTrue(),
